refactor(scripts): extract parsePrNumberFromUrl into shared lib/github-url.js (refs #3649)
Three inline copies of the PR-number-from-URL parser existed in finalizer.js,
watcher.js, and code-review.js, with a regex-anchoring inconsistency between
the watcher copy (^https://github.com/... anchor) and the other two (path-
segment match). Consolidates into a single parsePrNumberFromUrl export in
lib/github-url.js using the robust path-segment regex. Each original export
(extractPrNumber in finalizer/watcher, parsePrNumber in code-review) is
preserved as a re-export for backward compatibility with existing call sites
and tests.
